add mouse example
An interesting point to note is that up/down between mouse movement and
scrolling have opposite signs.

Another interesting point is that the mouse movements were not
registering with the 2s sleeps that were used in the virtual keyboard
example, so a shorter sleep is used. Perhaps my desktop environment is,
for reasons unknown to me, ignoring the first mouse event after a
certain period of no mouse events.

Co-authored-by: Leonid Satanovsky <leonid.satanovsky+leosat@gmail.com>